static cast "vec.size()" to "int" in ASSERT(start_pos < vec.size());

+++ tree/node.h
@@ -160,7 +160,7 @@ public:
         @param start_pos starting position in vec to copy to
     */
     virtual void getLength(DoubleVector &vec, int start_pos) { 
-        ASSERT(start_pos < vec.size());
+        ASSERT(start_pos < static_cast<int>(vec.size()));
         vec[start_pos] = length;
     }
 
@@ -197,7 +197,7 @@ public:
         @param start_pos starting position in vec to copy from
     */
     virtual void setLength(DoubleVector &vec, int start_pos, int num_elem) { 
-        ASSERT(start_pos < vec.size());
+        ASSERT(start_pos < static_cast<int>(vec.size()));
         ASSERT(num_elem == 1);
         length = vec[start_pos];
     }
